caching plantgrid too
