+++ plant-app/src/context/Plants.tsx
@@ -1,8 +1,9 @@
-import React, { createContext, useState, useContext } from "react";
+import React, { createContext, useState, useContext, useEffect } from "react";
 import { BASE_API_URL } from "../constants";
-import { Plant } from "../types/interfaces";
+import { Plant, PlantImage } from "../types/interfaces";
 import { useAlert } from "./Alerts";
 import { useApi } from "../utils/api";
+import noimagePlaceholder from "../assets/200x200_image_placeholder.png";
 
 interface PlantContextType {
   plants: Plant[];
@@ -10,6 +11,8 @@ interface PlantContextType {
   fetchPlants: (queryID: string) => Promise<void>;
   nextPlantId: number;
   forceReloadPlants: () => void;
+  plantImages: Record<string, string>;
+  plantGridIsLoading: boolean;
 }
 
 export const PlantContext = createContext<PlantContextType>(
@@ -38,6 +41,10 @@ export const PlantProvider: React.FC<PlantProviderProps> = ({ children }) => {
   const [isLoading, setIsLoading] = useState(true);
   const [nextPlantId, setNextPlantId] = useState(0);
   const [lastQueryID, setLastQueryID] = useState<string | null>(null);
+
+  const [plantGridIsLoading, setPlantGridIsLoading] = useState<boolean>(false);
+  const [plantImages, setPlantImages] = useState<Record<string, string>>({});
+
   const { showAlert } = useAlert();
   const { callApi } = useApi();
 
@@ -76,6 +83,43 @@ export const PlantProvider: React.FC<PlantProviderProps> = ({ children }) => {
       });
   };
 
+  // Fetch plant thumbnails for the grid
+  useEffect(() => {
+    // Check if image data for all plants already exists
+    if (!plants || plants.length === 0) {
+      return; // No need to fetch data
+    }
+    setPlantGridIsLoading(true);
+    const plantIds = plants.map((plant) => plant.plant_id);
+    console.log(plantIds);
+    callApi(BASE_API_URL + "/images/plants/most_recent/", {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify(plantIds),
+    }).then(async (response) => {
+      console.log(response);
+      if (!response.ok) {
+        showAlert("Error loading plant images", "danger");
+        return;
+      }
+      const imageData = await response.json();
+      const imageMap: Record<string, string> = {};
+      plantIds.forEach((id) => {
+        const foundImage = imageData.find(
+          (img: PlantImage) => img.plant_id === id,
+        );
+        imageMap[id] = foundImage
+          ? foundImage.signed_thumbnail_photo_url
+          : noimagePlaceholder;
+      });
+      setPlantImages(imageMap);
+      setPlantGridIsLoading(false);
+      showAlert("Loaded plant images", "success");
+    });
+  }, [plants, callApi, showAlert]);
+
   // TODO: convert this to force reload YOUR plants
   const forceReloadPlants = async () => {
     fetchPlants(lastQueryID || "", true);
@@ -83,11 +127,17 @@ export const PlantProvider: React.FC<PlantProviderProps> = ({ children }) => {
 
   // TODO: centralize all the plant create, update, delete methods here
 
-  // TODO: also add in method to fetch the PlantGrid images
-
   return (
     <PlantContext.Provider
-      value={{ plants, isLoading, fetchPlants, nextPlantId, forceReloadPlants }}
+      value={{
+        plants,
+        isLoading,
+        fetchPlants,
+        nextPlantId,
+        forceReloadPlants,
+        plantImages,
+        plantGridIsLoading,
+      }}
     >
       {children}
     </PlantContext.Provider>
